Reverts a14d2873ca488f430a2fba810740b525b9e8491f:
which states "Update the included COPYING file to the GPLv3. As stated in the README, the code is still 'GPL version 2 or later' but we want to encourage distribution under the GPLv3, so we include that version of the license for reference."
This fork has no plans to ever exercise the "later version" option.